rsx/vk: Bug fixes - Make each frame context own its own memory - Fix GPU blit - Fix image layout transitions in flip
vk: Improve frame-local memory usage tracking to prevent overwrites
- Also slightly bumps VRAM requirements for stream buffers to help with running out of storage
- Fixes flickering and missing graphics in some cases. Flickering is still there and needs more work
vk: Up vertex attribute heap size and increase the guard size on it
vulkan: Reorganize memory management
vulkan: blit cleanup
vulkan: blit engine improvements
- Override existing image mapping when conflicts detected
- Allow blitting of depth/stencil surfaces

+++ rpcs3/Emu/RSX/D3D12/D3D12RenderTargetSets.h
@@ -59,10 +59,10 @@ struct render_target_traits
 	{
 		//TODO
 		auto desc = surface->GetDesc();
-		info->rsx_pitch = desc.Width;
-		info->native_pitch = desc.Width;
-		info->surface_width = desc.Width;
-		info->surface_height = desc.Height;
+		info->rsx_pitch = static_cast<u16>(desc.Width);
+		info->native_pitch = static_cast<u16>(desc.Width);
+		info->surface_width = static_cast<u32>(desc.Width);
+		info->surface_height = static_cast<u32>(desc.Height);
 		info->bpp = 1;
 	}
